requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt
#
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
black==25.1.0
    # via eodhp-utils (pyproject.toml)
boto3==1.38.10
    # via
    #   eodhp-utils (pyproject.toml)
    #   moto
botocore==1.38.10
    # via
    #   boto3
    #   eodhp-utils (pyproject.toml)
    #   moto
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
certifi==2025.4.26
    # via
    #   pulsar-client
    #   requests
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   black
    #   pip-tools
coverage==7.8.0
    # via eodhp-utils (pyproject.toml)
cryptography==44.0.3
    # via moto
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
distlib==0.3.9
    # via virtualenv
execnet==2.1.1
    # via pytest-xdist
faker==37.1.0
    # via eodhp-utils (pyproject.toml)
fastjsonschema==2.21.1
    # via validate-pyproject
filelock==3.18.0
    # via virtualenv
identify==2.6.10
    # via pre-commit
idna==3.10
    # via requests
importlib-metadata==8.6.1
    # via opentelemetry-api
iniconfig==2.1.0
    # via pytest
isort==6.0.1
    # via eodhp-utils (pyproject.toml)
jinja2==3.1.6
    # via moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via eodhp-utils (pyproject.toml)
jsonschema-specifications==2025.4.1
    # via jsonschema
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
moto==5.1.4
    # via eodhp-utils (pyproject.toml)
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
opentelemetry-api==1.32.1
    # via
    #   eodhp-utils (pyproject.toml)
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-logging
    #   opentelemetry-processor-baggage
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.53b1
    # via opentelemetry-instrumentation-logging
opentelemetry-instrumentation-logging==0.53b1
    # via eodhp-utils (pyproject.toml)
opentelemetry-processor-baggage==0.53b1
    # via eodhp-utils (pyproject.toml)
opentelemetry-sdk==1.32.1
    # via
    #   eodhp-utils (pyproject.toml)
    #   opentelemetry-processor-baggage
opentelemetry-semantic-conventions==0.53b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==25.0
    # via
    #   black
    #   build
    #   opentelemetry-instrumentation
    #   pytest
    #   validate-pyproject
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via eodhp-utils (pyproject.toml)
platformdirs==4.3.7
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.2.0
    # via eodhp-utils (pyproject.toml)
pulsar-client==3.6.1
    # via eodhp-utils (pyproject.toml)
pycparser==2.22
    # via cffi
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pysubnettree==0.37
    # via eodhp-utils (pyproject.toml)
pytest==8.3.5
    # via
    #   eodhp-utils (pyproject.toml)
    #   pytest-mock
    #   pytest-xdist
pytest-mock==3.14.0
    # via eodhp-utils (pyproject.toml)
pytest-watcher==0.4.3
    # via eodhp-utils (pyproject.toml)
pytest-xdist==3.6.1
    # via eodhp-utils (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   moto
python-json-logger==3.3.0
    # via eodhp-utils (pyproject.toml)
pyyaml==6.0.2
    # via
    #   pre-commit
    #   responses
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   eodhp-utils (pyproject.toml)
    #   moto
    #   requests-mock
    #   responses
requests-mock==1.12.1
    # via eodhp-utils (pyproject.toml)
responses==0.25.7
    # via moto
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
ruff==0.11.8
    # via eodhp-utils (pyproject.toml)
s3transfer==0.12.0
    # via boto3
six==1.17.0
    # via python-dateutil
trove-classifiers==2025.5.1.12
    # via validate-pyproject
typing-extensions==4.13.2
    # via
    #   opentelemetry-sdk
    #   referencing
tzdata==2025.2
    # via faker
urllib3==2.4.0
    # via
    #   botocore
    #   requests
    #   responses
validate-pyproject[all]==0.24.1
    # via eodhp-utils (pyproject.toml)
virtualenv==20.31.1
    # via pre-commit
watchdog==6.0.0
    # via pytest-watcher
werkzeug==3.1.3
    # via moto
wheel==0.45.1
    # via pip-tools
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-processor-baggage
xmltodict==0.14.2
    # via moto
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
